breaks up base class
