+++ scripts/walk_through_session.py
@@ -2,7 +2,7 @@ import logging
 import os
 
 from aind_behavior_dynamic_foraging.data_contract import dataset as df_foraging_dataset
-from aind_behavior_dynamic_foraging.task_logic.trial_generators.warmup_trial_generator import WarmupTrialGeneratorSpec
+from aind_behavior_dynamic_foraging.task_logic.trial_generators.coupled_warmup_trial_generator import CoupledWarmupTrialGeneratorSpec
 from aind_behavior_dynamic_foraging.task_logic.trial_models import TrialOutcome
 
 logging.basicConfig(
@@ -17,7 +17,7 @@ def walk_through_session(data_directory: os.PathLike):
     software_events.load_all()
 
     trial_outcomes = software_events["TrialOutcome"].data["data"].iloc
-    warmup_trial_generator = WarmupTrialGeneratorSpec().create_generator()
+    warmup_trial_generator = CoupledWarmupTrialGeneratorSpec().create_generator()
     for i, outcome in enumerate(trial_outcomes):
         warmup_trial_generator.update(TrialOutcome.model_validate(outcome))
         trial = warmup_trial_generator.next()
