+++ src/aind_behavior_dynamic_foraging/task_logic/trial_generators/__init__.py
@@ -7,12 +7,12 @@ from ._base import ITrialGenerator as ITrialGenerator
 from .composite_trial_generator import TrialGeneratorCompositeSpec
 from .coupled_trial_generator import CoupledTrialGeneratorSpec
 from .integration_test_trial_generator import IntegrationTestTrialGeneratorSpec
-from .warmup_trial_generator import WarmupTrialGeneratorSpec
+from .coupled_warmup_trial_generator import CoupledWarmupTrialGeneratorSpec
 
 if TYPE_CHECKING:
     TrialGeneratorSpec = Union[
         (
-            WarmupTrialGeneratorSpec,
+            CoupledWarmupTrialGeneratorSpec,
             CoupledTrialGeneratorSpec,
             IntegrationTestTrialGeneratorSpec,
             TrialGeneratorCompositeSpec["TrialGeneratorSpec"],
@@ -24,7 +24,7 @@ else:
         Annotated[
             Union[
                 (
-                    WarmupTrialGeneratorSpec,
+                    CoupledWarmupTrialGeneratorSpec,
                     CoupledTrialGeneratorSpec,
                     IntegrationTestTrialGeneratorSpec,
                     TrialGeneratorCompositeSpec["TrialGeneratorSpec"],
