+++ src/aind_behavior_dynamic_foraging/task_logic/trial_generators/block_based_trial_generator.py
@@ -20,36 +20,14 @@ from ._base import BaseTrialGeneratorSpecModel, ITrialGenerator
 logger = logging.getLogger(__name__)
 
 
-class RewardProbabilityParameters(BaseModel):
-    """Defines the reward probability structure for a dynamic foraging task.
-
-    Reward probabilities are defined as pairs (p_left, p_right) normalized by
-    base_reward_sum. Pairs are drawn from a family representing a difficulty level:
-
-        Family 1:   [[8, 1], [6, 1], [3, 1], [1, 1]]
-        Family 2:  [[8, 1], [1, 1]]
-        Family 3:  [[1.0, 0.0], [0.9, 0.1], [0.8, 0.2], [0.7, 0.3], [0.6, 0.4], [0.5, 0.5]]
-        Family 4:  [[6, 1], [3, 1], [1, 1]]
-
-    """
-
-    base_reward_sum: float = Field(
-        default=0.8,
-        description="Total reward probability shared between the two sides. Each reward pair is normalized to sum to this value.",
-    )
-    reward_pairs: list[list[float, float]] = Field(
-        default=[[8, 1]],
-        description="List of (left, right) reward ratio pairs to sample from during block transitions. ",
-    )
-
 
 class Block(BaseModel):
-    p_right_reward: float = Field(ge=0, le=1, description="Reward probability for right side during block.")
-    p_left_reward: float = Field(ge=0, le=1, description="Reward probability for left side during block.")
+    p_right_reward: Optional[float] = Field(ge=0, le=1, description="Reward probability for right side during block.")
+    p_left_reward: Optional[float] = Field(ge=0, le=1, description="Reward probability for left side during block.")
     min_length: int = Field(ge=0, description="Minimum number of trials in block.")
 
 
-class BlockBasedTrialGeneratorSpec(BaseTrialGeneratorSpecModel):
+class TwoArmedBanditTrialGeneratorSpec(BaseTrialGeneratorSpecModel):
     type: Literal["BlockBasedTrialGenerator"] = "BlockBasedTrialGenerator"
 
     quiescent_duration: Distribution = Field(
@@ -84,22 +62,13 @@ class BlockBasedTrialGeneratorSpec(BaseTrialGeneratorSpecModel):
         description="Distribution describing block length.",
     )
 
-    min_block_reward: int = Field(default=1, ge=0, title="Minimal rewards in a block to switch")
-
-    kernel_size: int = Field(default=2, description="Kernel to evaluate choice fraction.")
-    reward_probability_parameters: RewardProbabilityParameters = Field(
-        default=RewardProbabilityParameters(),
-        description="Parameters defining the reward probability structure.",
-        validate_default=True,
-    )
-
     is_baiting: bool = Field(default=False, description="Whether uncollected rewards carry over to the next trial.")
 
     def create_generator(self) -> "BlockBasedTrialGenerator":
         return BlockBasedTrialGenerator(self)
 
 
-class BlockBasedTrialGenerator(ITrialGenerator, ABC):
+class TwoArmedBanditTrialGenerator(ITrialGenerator, ABC):
     """Abstract trial generator for block-based dynamic foraging tasks.
 
     Manages block transitions, baiting logic, and trial generation. Subclasses
@@ -117,25 +86,22 @@ class BlockBasedTrialGenerator(ITrialGenerator, ABC):
         is_right_baited: Whether the right port currently has a baited reward.
     """
 
-    def __init__(self, spec: BlockBasedTrialGeneratorSpec) -> None:
+    def __init__(self, spec: TwoArmedBanditTrialGeneratorSpec) -> None:
         """Initializes the generator and generates the first block.
 
         Args:
-            spec: The BlockBasedTrialGeneratorSpec defining task parameters.
+            spec: The TwoArmedBanditTrialGeneratorSpec defining task parameters.
         """
 
         self.spec = spec
         self.is_right_choice_history: list[bool | None] = []
         self.reward_history: list[bool] = []
-        self.block_history: list[Block] = []
-        self.block: Block = self._generate_next_block(
-            reward_pairs=self.spec.reward_probability_parameters.reward_pairs,
-            base_reward_sum=self.spec.reward_probability_parameters.base_reward_sum,
-            block_len=self.spec.block_len,
-        )
         self.trials_in_block = 0
         self.is_left_baited: bool = False
         self.is_right_baited: bool = False
+        self.p_left_reward: float
+        self.p_right_reward: float
+        self.block: Block
 
     def next(self) -> Trial | None:
         """Generates the next trial in the session.
@@ -157,29 +123,25 @@ class BlockBasedTrialGenerator(ITrialGenerator, ABC):
         iti = draw_sample(self.spec.inter_trial_interval_duration)
         quiescent = draw_sample(self.spec.quiescent_duration)
 
-        p_reward_left = self.block.p_left_reward
-        p_reward_right = self.block.p_right_reward
-
         if self.spec.is_baiting:
             random_numbers = np.random.random(2)
 
-            is_left_baited = self.block.p_left_reward > random_numbers[0] or self.is_left_baited
-            logger.debug(f"Left baited: {is_left_baited}")
-            p_reward_left = 1 if is_left_baited else p_reward_left
+            self.is_left_baited = self.p_left_reward > random_numbers[0] or self.is_left_baited
+            logger.debug(f"Left baited: {self.is_left_baited}")
 
-            is_right_baited = self.block.p_right_reward > random_numbers[1] or self.is_right_baited
-            logger.debug(f"Right baited: {is_left_baited}")
-            p_reward_right = 1 if is_right_baited else p_reward_right
+            self.is_right_baited = self.p_right_reward > random_numbers[1] or self.is_right_baited
+            logger.debug(f"Right baited: {self.is_right_baited}")
+
 
         return Trial(
-            p_reward_left=p_reward_left,
-            p_reward_right=p_reward_right,
+            p_reward_left=1 if (self.is_left_baited and self.spec.is_baiting) else self.p_left_reward,
+            p_reward_right=1 if (self.is_right_baited and self.spec.is_baiting) else self.p_right_reward,
             reward_consumption_duration=self.spec.reward_consumption_duration,
             response_deadline_duration=self.spec.response_duration,
             quiescence_period_duration=quiescent,
             inter_trial_interval_duration=iti,
         )
-
+    
     @abstractmethod
     def _are_end_conditions_met(self) -> bool:
         """Checks whether the session should end.
@@ -189,71 +151,25 @@ class BlockBasedTrialGenerator(ITrialGenerator, ABC):
             generated, False otherwise.
         """
         pass
-
-    def _generate_next_block(
-        self,
-        reward_pairs: list[list[float, float]],
-        base_reward_sum: float,
-        block_len: Union[UniformDistribution, ExponentialDistribution],
-        current_block: Optional[Block] = None,
-    ) -> Block:
-        """Generates the next block, avoiding repeating the current block's side bias.
-
-        Normalizes reward pairs by base_reward_sum, mirrors them to create a full
-        pool, optionally excludes the current block's probabilities and high-reward
-        side, then randomly samples the next block.
-
-        Args:
-            reward_pairs: List of (left, right) reward ratio pairs to draw from.
-            base_reward_sum: Total reward probability to normalize each pair to.
-            block_len: Distribution from which to sample the next block length.
-            current_block: The currently active block, used to avoid repeating the
-                same reward probabilities or high-reward side. Defaults to None.
+    
+    @staticmethod
+    @abstractmethod
+    def _generate_next_block(*args, **kwargs) -> Block:
+        """Abstract method. Subclasses must implement their own block switching logic.
 
         Returns:
             A new Block with sampled reward probabilities and length.
         """
 
-        logger.info("Generating next block.")
+        pass
 
-        # determine candidate reward pairs
-        reward_prob = np.array(reward_pairs, dtype=float)
-        reward_prob /= reward_prob.sum(axis=1, keepdims=True)
-        reward_prob *= float(base_reward_sum)
-        logger.info(f"Candidate reward pairs normalized and scaled: {reward_prob.tolist()}")
+    @abstractmethod
+    def _is_block_switch_allowed(self) -> bool:
 
-        # create pool including all reward probabiliteis and mirrored pairs
-        reward_prob_pool = np.vstack([reward_prob, np.fliplr(reward_prob)])
+        """Determines whether all criteria are met to switch to the next block.
 
-        if current_block:  # exclude previous block if history exists
-            logger.info("Excluding previous block reward probability.")
-            last_block_reward_prob = [current_block.p_right_reward, current_block.p_left_reward]
+        Returns:
+            True if all switch criteria are satisfied, False otherwise.
+        """
 
-            # remove blocks identical to last block
-            reward_prob_pool = reward_prob_pool[np.any(reward_prob_pool != last_block_reward_prob, axis=1)]
-            logger.debug(f"Pool after removing identical to last block: {reward_prob_pool.tolist()}")
-
-            # remove blocks with same high-reward side (if last block had a clear high side)
-            if last_block_reward_prob[0] != last_block_reward_prob[1]:
-                high_side_last = last_block_reward_prob[0] > last_block_reward_prob[1]
-                high_side_pool = reward_prob_pool[:, 0] > reward_prob_pool[:, 1]
-                reward_prob_pool = reward_prob_pool[high_side_pool != high_side_last]
-                logger.debug(f"Pool after removing same high-reward side: {reward_prob_pool.tolist()}")
-
-        # remove duplicates
-        reward_prob_pool = np.unique(reward_prob_pool, axis=0)
-        logger.debug(f"Final reward probability pool after removing duplicates: {reward_prob_pool.tolist()}")
-
-        # randomly pick next block reward probability
-        p_right_reward, p_left_reward = reward_prob_pool[random.choice(range(reward_prob_pool.shape[0]))]
-        logger.info(f"Selected next block reward probabilities: right={p_right_reward}, left={p_left_reward}")
-
-        # randomly pick block length
-        next_block_len = round(draw_sample(block_len))
-        logger.info(f"Selected next block length: {next_block_len}")
-
-        return Block(
-            p_right_reward=p_right_reward,
-            p_left_reward=p_left_reward,
-            min_length=next_block_len,
-        )
+        pass
