+++ src/aind_behavior_dynamic_foraging/task_logic/trial_generators/coupled_trial_generator.py
@@ -6,9 +6,9 @@ import numpy as np
 from pydantic import BaseModel, Field
 
 from ..trial_models import TrialOutcome
-from .block_based_trial_generator import (
-    BlockBasedTrialGenerator,
-    BlockBasedTrialGeneratorSpec,
+from .coupled_warmup_trial_generator import (
+    CoupledWarmupTrialGenerator,
+    CoupledWarmupTrialGeneratorSpec,
 )
 
 logger = logging.getLogger(__name__)
@@ -55,7 +55,7 @@ class BehaviorStabilityParameters(BaseModel):
     )
 
 
-class CoupledTrialGeneratorSpec(BlockBasedTrialGeneratorSpec):
+class CoupledTrialGeneratorSpec(CoupledWarmupTrialGeneratorSpec):
     type: Literal["CoupledTrialGenerator"] = "CoupledTrialGenerator"
 
     trial_generation_end_parameters: CoupledTrialGenerationEndConditions = Field(
@@ -73,11 +73,15 @@ class CoupledTrialGeneratorSpec(BlockBasedTrialGeneratorSpec):
         description="Whether to extend the minimum block length by one trial when the animal does not respond.",
     )
 
+    min_block_reward: int = Field(default=1, ge=0, title="Minimal rewards in a block to switch")
+
+    kernel_size: int = Field(default=2, description="Kernel to evaluate choice fraction.")
+
     def create_generator(self) -> "CoupledTrialGenerator":
         return CoupledTrialGenerator(self)
 
 
-class CoupledTrialGenerator(BlockBasedTrialGenerator):
+class CoupledTrialGenerator(CoupledWarmupTrialGenerator):
     """Trial generator for a coupled block-based dynamic foraging task.
 
     Extends BlockBasedTrialGenerator with session end conditions, baiting state
@@ -135,60 +139,6 @@ class CoupledTrialGenerator(BlockBasedTrialGenerator):
             return True
 
         return False
-
-    def update(self, outcome: TrialOutcome | str) -> None:
-        """Updates generator state from the previous trial outcome and switches block if criteria are met.
-
-        Records choice and reward history, manages baiting state, optionally extends
-        the block on no response, and triggers a block switch if all switch criteria
-        are satisfied.
-
-        Args:
-            outcome: The TrialOutcome from the most recently completed trial.
-        """
-
-        logger.info(f"Updating coupled trial generator with trial outcome of {outcome}")
-
-        if isinstance(outcome, str):
-            outcome = TrialOutcome.model_validate_json(outcome)
-
-        self.is_right_choice_history.append(outcome.is_right_choice)
-        self.reward_history.append(outcome.is_rewarded)
-        self.trials_in_block += 1
-
-        if self.spec.is_baiting:
-            if outcome.is_right_choice:
-                logger.debug("Resesting right bait.")
-                self.is_right_baited = False
-            elif outcome.is_right_choice is False:
-                logger.debug("Resesting left bait.")
-                self.is_left_baited = False
-
-        if self.spec.extend_block_on_no_response and outcome.is_right_choice is None:
-            logger.info("Extending minimum block length due to ignored trial.")
-            self.block.min_length += 1
-
-        switch_block = self._is_block_switch_allowed(
-            trials_in_block=self.trials_in_block,
-            min_block_reward=self.spec.min_block_reward,
-            choice_history=self.is_right_choice_history,
-            p_right_reward=self.block.p_right_reward,
-            p_left_reward=self.block.p_left_reward,
-            beh_stability_params=self.spec.behavior_stability_parameters,
-            block_length=self.block.min_length,
-            kernel_size=self.spec.kernel_size,
-        )
-
-        if switch_block:
-            logger.info("Switching block.")
-            self.trials_in_block = 0
-            self.block = self._generate_next_block(
-                reward_pairs=self.spec.reward_probability_parameters.reward_pairs,
-                base_reward_sum=self.spec.reward_probability_parameters.base_reward_sum,
-                current_block=self.block,
-                block_len=self.spec.block_len,
-            )
-            self.block_history.append(self.block)
 
     def _is_behavior_stable(
         self,
@@ -306,33 +256,12 @@ class CoupledTrialGenerator(BlockBasedTrialGenerator):
             choice_fraction[i] = np.nanmean(window)
         return choice_fraction
 
-    def _is_block_switch_allowed(
-        self,
-        trials_in_block: int,
-        min_block_reward: int,
-        choice_history: list,
-        p_right_reward: float,
-        p_left_reward: float,
-        beh_stability_params: BehaviorStabilityParameters,
-        block_length: int,
-        kernel_size: int = 2,
-    ) -> bool:
+    def _is_block_switch_allowed(self) -> bool:
         """Determines whether all criteria are met to switch to the next block.
 
         A block switch requires: the planned block length has been reached, the
         minimum reward count has been collected, and behavior meets the stability
         criterion.
-
-        Args:
-            trials_in_block: Number of trials elapsed in the current block.
-            min_block_reward: Minimum total rewards required before switching.
-            choice_history: Trial history with True for right, False for left,
-                and None for ignored trials.
-            p_right_reward: Reward probability for the right port in the current block.
-            p_left_reward: Reward probability for the left port in the current block.
-            beh_stability_params: Parameters defining the behavior stability criterion.
-            block_length: Planned minimum number of trials in the current block.
-            kernel_size: Sliding window size for computing choice fraction.
 
         Returns:
             True if all switch criteria are satisfied, False otherwise.
@@ -340,23 +269,27 @@ class CoupledTrialGenerator(BlockBasedTrialGenerator):
 
         logger.info("Evaluating block switch.")
 
+        if self.spec.extend_block_on_no_response and self.is_right_choice_history[-1] is None:
+            logger.info("Extending minimum block length due to ignored trial.")
+            self.block.min_length += 1
+
         # has planned block length been reached?
-        block_length_ok = trials_in_block >= block_length
+        block_length_ok = self.trials_in_block >= self.block.min_length
         logger.debug(f"Planned block length reached: {block_length_ok}")
 
         # is behavior qualified to switch?
         behavior_ok = self._is_behavior_stable(
-            choice_history,
-            p_right_reward,
-            p_left_reward,
-            beh_stability_params,
-            trials_in_block,
-            kernel_size,
+            self.is_right_choice_history,
+            self.block.p_right_reward,
+            self.block.p_left_reward,
+            self.spec.behavior_stability_parameters,
+            self.trials_in_block,
+            self.spec.kernel_size,
         )
         logger.debug(f"Behavior meets stability criteria: {behavior_ok}")
 
         # has reward criteria been met?
-        reward_ok = self.reward_history.count(False) + self.reward_history.count(True) >= min_block_reward
+        reward_ok = self.reward_history.count(False) + self.reward_history.count(True) >= self.spec.min_block_reward
         logger.debug(f"Reward criterion satisfied: {reward_ok}")
 
         # conditions to switch:
